refactor: move source folders into src/
