+++ cypress/component/ui/Button.cy.tsx
@@ -1,4 +1,4 @@
-import Button from '../../../components/ui/Button';
+import Button from '../../../src/components/ui/Button';
 
 describe('Button.cy.ts', () => {
   it('calls the onClick callback when clicked', () => {
