+++ sonar-project.properties
@@ -1,6 +1,6 @@
 sonar.projectKey=antwika_web
 sonar.organization=antwika
-sonar.sources=pages,components,context,hooks,misc,redux
+sonar.sources=src
 sonar.exclusions=node_modules/**
 sonar.test.inclusions=test/**
 sonar.javascript.lcov.reportPaths=coverage/lcov.info
