+++ src/components/ExpandingBox.module.css
