+++ src/components/ExpandingBox.tsx
