+++ src/components/TypedText.tsx
