+++ src/hooks/useTypedText.ts
