+++ src/pages/_app.tsx
@@ -1,4 +1,4 @@
-import '../styles/globals.css'
+import '../../styles/globals.css'
 import type { AppProps } from 'next/app'
 import { IntlProvider } from 'react-intl'
 import { useRouter } from 'next/router';
